refactor: NewsViewModel 위치 변경
